+++ unrest/tests/flask/test_post.py
@@ -152,7 +152,7 @@ def test_post_with_defaults(rest, http):
         methods=['GET', 'POST'],
         defaults={
             'color': 'white',
-            'age': 42.0,
+            'age': lambda p: p['size'] * 2,
         }
     )
     code, json = http.post('/api/fruit', json={'size': 1.0, 'tree_id': 1})
@@ -163,7 +163,7 @@ def test_post_with_defaults(rest, http):
             'fruit_id': 6,
             'color': 'white',
             'size': 1.0,
-            'age': 42.0,
+            'age': 2.0,
             'tree_id': 1
         },
     ]
@@ -181,7 +181,7 @@ def test_post_with_defaults(rest, http):
             'fruit_id': 7,
             'color': 'yellow',
             'size': 2.0,
-            'age': 42.0,
+            'age': 4.0,
             'tree_id': 2
         },
     ]
@@ -194,14 +194,14 @@ def test_post_with_defaults(rest, http):
             'fruit_id': 6,
             'color': 'white',
             'size': 1.0,
-            'age': 42.0,
+            'age': 2.0,
             'tree_id': 1
         },
         {
             'fruit_id': 7,
             'color': 'yellow',
             'size': 2.0,
-            'age': 42.0,
+            'age': 4.0,
             'tree_id': 2
         },
     ]
@@ -213,7 +213,7 @@ def test_post_with_fixed(rest, http):
         methods=['GET', 'POST'],
         fixed={
             'color': 'white',
-            'age': 42.0,
+            'age': lambda p: p['size'] * 2,
         }
     )
     code, json = http.post('/api/fruit', json={'size': 1.0, 'tree_id': 1})
@@ -224,7 +224,7 @@ def test_post_with_fixed(rest, http):
             'fruit_id': 6,
             'color': 'white',
             'size': 1.0,
-            'age': 42.0,
+            'age': 2.0,
             'tree_id': 1
         },
     ]
@@ -242,7 +242,7 @@ def test_post_with_fixed(rest, http):
             'fruit_id': 7,
             'color': 'white',
             'size': 2.0,
-            'age': 42.0,
+            'age': 4.0,
             'tree_id': 2
         },
     ]
@@ -255,14 +255,14 @@ def test_post_with_fixed(rest, http):
             'fruit_id': 6,
             'color': 'white',
             'size': 1.0,
-            'age': 42.0,
+            'age': 2.0,
             'tree_id': 1
         },
         {
             'fruit_id': 7,
             'color': 'white',
             'size': 2.0,
-            'age': 42.0,
+            'age': 4.0,
             'tree_id': 2
         },
     ]